repo

+++ Dtos/ProductDto.cs
@@ -4,14 +4,14 @@ namespace ProductApi.Dto
     {
         public int Id { get; set; }
         public string Name { get; set; } = "";
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public string Description { get; set; } = "";
         public int CategoryId { get; set; }
     }
     public class CreateProductDto
     {
         public string Name { get; set; } = "";
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public string Description { get; set; } = "";
         public int CategoryId { get; set; }
     }
@@ -19,7 +19,7 @@ namespace ProductApi.Dto
     {
         public int Id { get; set; }
         public string Name { get; set; } = "";
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public string Description { get; set; } = "";
         public int CategoryId { get; set; }
     }
